moved SalesforceAuthenticate (and accompanying tests) into 'lib' to make available for re-use
