+++ handlers/identity-backfill/src/test/scala/com/gu/identityBackfill/salesforce/updateSFIdentityId/UpdateSalesforceIdentityIdTest.scala
@@ -3,8 +3,9 @@ package com.gu.identityBackfill.salesforce.updateSFIdentityId
 import com.gu.effects.TestingRawEffects
 import com.gu.effects.TestingRawEffects.{HTTPResponse, POSTRequest}
 import com.gu.identityBackfill.Types.{IdentityId, SFContactId}
-import com.gu.identityBackfill.salesforce.SalesforceAuthenticate.SalesforceAuth
-import com.gu.identityBackfill.salesforce.{SalesforceRestRequestMaker, UpdateSalesforceIdentityId}
+import com.gu.salesforce.auth.SalesforceAuthenticate.SalesforceAuth
+import com.gu.identityBackfill.salesforce.UpdateSalesforceIdentityId
+import com.gu.salesforce.auth.SalesforceRestRequestMaker
 import org.scalatest.{FlatSpec, Matchers}
 import scalaz.\/-
